requirements.txt
pyrogram
tgcrypto
aiohttp
flask
yt-dlp
moviepy
pillow
requests